Edit: moved choose game providers to respective folders in domain and infra layers
