--- src/modules/chooseGame/container/providers/CryptoProvider/index.ts
@@ -1,6 +1,0 @@
-import CryptoJDCryptoProvider from 'modules/chooseGame/container/providers/CryptoProvider/implementations/CryptoJSCryptoProvider';
-import ICryptoProvider from 'modules/chooseGame/container/providers/CryptoProvider/models/ICryptoProvider';
-
-export default function makeCryptoProvider(): ICryptoProvider {
-  return new CryptoJDCryptoProvider();
-}
