--- src/modules/chooseGame/container/providers/index.ts
@@ -1,1 +1,0 @@
-export { default as makeCryptoProvider } from './CryptoProvider';
