+++ src/modules/chooseGame/domain/providers/ICryptoProvider.ts
