+++ src/modules/chooseGame/infra/providers/factories/makeCryptoProvider.ts
@@ -1,0 +1,6 @@
+import ICryptoProvider from 'modules/chooseGame/domain/providers/ICryptoProvider';
+import CryptoJDCryptoProvider from 'modules/chooseGame/infra/providers/CryptoJSCryptoProvider';
+
+export default function makeCryptoProvider(): ICryptoProvider {
+  return new CryptoJDCryptoProvider();
+}
